+++ app/src/main/java/com/example/friendlygram/ui/fragments/BaseChangeFragment.kt
@@ -5,6 +5,7 @@ import androidx.fragment.app.Fragment
 import com.example.friendlygram.MainActivity
 import com.example.friendlygram.R
 import com.example.friendlygram.utitits.APP_ACTIVITY
+import com.example.friendlygram.utitits.hideKeyboard
 
 
 open class BaseChangeFragment(layout:Int) : Fragment(layout) {
@@ -14,13 +15,12 @@ open class BaseChangeFragment(layout:Int) : Fragment(layout) {
         setHasOptionsMenu(true)
         super.onStart()
         (APP_ACTIVITY).mAppDrawer.disableDrawer()
+        hideKeyboard()
     }
 
     override fun onStop() {
         super.onStop()
-        APP_ACTIVITY.hideKeyboard()
-
-    }
+            }
 
     override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
         (APP_ACTIVITY).menuInflater.inflate(R.menu.settings_menu_confirm, menu)
