+++ app/src/main/java/com/example/friendlygram/ui/fragments/SettingsFragment.kt
@@ -8,6 +8,7 @@ import android.view.MenuItem
 import com.example.friendlygram.R
 import com.example.friendlygram.activities.RegisterActivity
 import com.example.friendlygram.utitits.*
+import com.squareup.picasso.Picasso
 import com.theartofdev.edmodo.cropper.CropImage
 import com.theartofdev.edmodo.cropper.CropImageView
 import kotlinx.android.synthetic.main.fragment_settings.*
@@ -41,8 +42,8 @@ class SettingsFragment : BaseFragment(R.layout.fragment_settings) {
             .setAspectRatio(1,1)
             .setRequestedSize(600,600)
             .setCropShape(CropImageView.CropShape.OVAL)
-            .start((APP_ACTIVITY))
-            .setResult(RESULT_OK)
+            .start(APP_ACTIVITY,this)
+
 
 
     }
@@ -65,7 +66,39 @@ class SettingsFragment : BaseFragment(R.layout.fragment_settings) {
         }
         return true
     }
+    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
+        super.onActivityResult(requestCode, resultCode, data)
+        if (requestCode== CropImage.CROP_IMAGE_ACTIVITY_REQUEST_CODE && resultCode == RESULT_OK && data != null){
+            val uri = CropImage.getActivityResult(data).uri
+            val path = REF_STORAGE_ROOT.child(FOLDER_PROFILE_IMAGE)
+                .child(CURRENT_UID)
+            path.putFile(uri).addOnCompleteListener { task1 ->
+                if (task1.isSuccessful){
+                    path.downloadUrl.addOnCompleteListener { task2 ->
+                        if (task2.isSuccessful){
+                            val  photoUrl = task2.result.toString()
+                            REF_DATABASE_ROOT.child(NODE_USERS).child(CURRENT_UID).child(
+                                CHILD_PHOTO_URL).setValue(photoUrl)
+                                .addOnCompleteListener {
+                                    if (it.isSuccessful){
+                                        settings_user_photo.downloadAndSetImage(photoUrl)
+                                      showToast(getString(R.string.toast_data_update))
+                                        USER.photoUrl = photoUrl
+                                    }
+
+                                }
+                        }
+                    }
+
+                }
+            }
+        }
+    }
+
+
 }
+
+
 
 private fun Any.setResult(resultOk: Int) {
