+++ app/src/main/java/com/example/friendlygram/utitits/appDatabaseHelper.kt
@@ -23,6 +23,8 @@ const val CHILD_PHONE = "phone"
 const val CHILD_USERNAME = "username"
 const val CHILD_FULLNAME = "fullname"
 const val CHILD_BIO = "bio"
+const val CHILD_PHOTO_URL ="photoUrl"
+
 
 fun initFirebase() {
     AUTH = FirebaseAuth.getInstance()
